requirements.txt
pandas==1.1.0
django==3.1
django-cors-headers==3.4.0
django-timezone-field==4.0
django-filter==2.3.0
django-redis==4.12.1
django-redis-cache==2.1.1
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
mysqlclient==2.0.1
drf-extensions==0.6.0
drf-yasg==1.17.1
django_notifications_hq==1.6.0
pycryptodome==3.9.8
